refactor: 헤더 프로필 드롭다운과 아바타 메뉴 로직 공통화
