+++ src/types/layout.ts
@@ -1,0 +1,6 @@
+export interface ProfileMenuItem {
+  id: string
+  label: string
+  onSelect: () => void
+  tone?: 'default' | 'danger'
+}
